+++ common/src/main/java/com/sonicether/soundphysics/integration/ClothConfigIntegration.java
@@ -92,6 +92,11 @@ public class ClothConfigIntegration {
                 Component.translatable("cloth_config.sound_physics_remastered.underwater_filter.description"),
                 SoundPhysicsMod.CONFIG.underwaterFilter
         ));
+        general.addEntry(fromConfigEntry(entryBuilder,
+                Component.translatable("cloth_config.sound_physics_remastered.evaluate_ambient_sounds"),
+                Component.translatable("cloth_config.sound_physics_remastered.evaluate_ambient_sounds.description"),
+                SoundPhysicsMod.CONFIG.evaluateAmbientSounds
+        ));
 
         ConfigCategory performance = builder.getOrCreateCategory(Component.translatable("cloth_config.sound_physics_remastered.category.performance"));