+++ fabric/changelog.md
@@ -3,3 +3,5 @@
 - Fixed underwater sounds
 - Fixed boat rowing sounds muffled
 - Fixed player hurt sound muffled when standing in water
+- Fixed not being able to hear ambient cave sounds
+- Fixed ambient underwater sounds being muffled